+++ packages/cli/src/commands/bundle.ts
@@ -36,6 +36,7 @@ export async function handleBundle(
   config.lint.skipDecorators(argv['skip-decorator']);
   const entrypoints = await getFallbackEntryPointsOrExit(argv.entrypoints, config);
   const totals: Totals = { errors: 0, warnings: 0, ignored: 0 };
+  const maxProblems = argv['max-problems'];
 
   for (const entrypoint of entrypoints) {
     try {
@@ -46,9 +47,13 @@ export async function handleBundle(
           ref: entrypoint,
           config,
         });
-
         const fileLintTotals = getTotals(results);
-        formatProblems(results, { format: 'stylish', totals: fileLintTotals, version });
+
+        totals.errors += fileLintTotals.errors;
+        totals.warnings += fileLintTotals.warnings;
+        totals.ignored += fileLintTotals.ignored;
+
+        formatProblems(results, { format: argv.format || 'codeframe', totals: fileLintTotals, version, maxProblems });
         printLintTotals(fileLintTotals, 2);
       }
 
@@ -83,7 +88,7 @@ export async function handleBundle(
 
       formatProblems(problems, {
         format: argv.format,
-        maxProblems: argv['max-problems'],
+        maxProblems,
         totals: fileTotals,
         version,
       });