+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-export { BundleOutputFormat, readFileFromUrl } from './utils';
+export { BundleOutputFormat, readFileFromUrl, slash } from './utils';
 export { Oas3_1Types } from './types/oas3_1';
 export { Oas3Types } from './types/oas3';
 export { Oas2Types } from './types/oas2';
@@ -29,9 +29,11 @@ export {
   YamlParseError,
   makeDocumentFromString,
 } from './resolve';
+export { parseYaml, stringifyYaml } from './js-yaml';
 export { unescapePointer } from './ref-utils';
 export { detectOpenAPI, OasMajorVersion, openAPIMajor, OasVersion } from './oas-types';
 export { normalizeVisitors } from './visitors';
+
 export {
   WalkContext,
   walkDocument,